// apps/frontend/src/api/models/createJobRequest.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { CreateJobRequestDescription } from './createJobRequestDescription';

export interface CreateJobRequest {
  title: string;
  company: string;
  application_url: string;
  description?: CreateJobRequestDescription;
}